+++ code/kdTree/parallelApproved/buildTree_serial.cpp
@@ -23,7 +23,7 @@ using namespace std;
 // Function: buildTree
 //
 
-void buildTree_serial(double *data, int *rows, int cols, Tree *tree) {
+void buildTree_serial(float *data, int *rows, int cols, Tree *tree) {
 
 	// function tree = buildTree( data, tree )
 	//
