+++ code/kdTree/parallelApproved/searchTree.cpp
@@ -28,7 +28,7 @@ using namespace std;
 // Function: buildTree
 //
 
-void searchTree( double *point, double rad, double *data[], int *rows, int cols, Tree *tree, MPI_Comm comm, int myRank, int numNodes, string name ) {
+void searchTree( float *point, float rad, float *data[], int *rows, int cols, Tree *tree, MPI_Comm comm, int myRank, int numNodes, string name ) {
 
 	int q = 0, currentRank = 0, rootRank;
 	MPI_Comm_size( comm, &q );
